Move DatabaseException to database
