+++ app/src/main/java/com/nervousfish/nervousfish/modules/database/GsonDatabaseAdapter.java
@@ -6,7 +6,6 @@ import com.nervousfish.nervousfish.data_objects.Contact;
 import com.nervousfish.nervousfish.data_objects.Database;
 import com.nervousfish.nervousfish.data_objects.IKey;
 import com.nervousfish.nervousfish.data_objects.Profile;
-import com.nervousfish.nervousfish.exceptions.DatabaseAdapterException;
 import com.nervousfish.nervousfish.modules.cryptography.IEncryptor;
 import com.nervousfish.nervousfish.modules.filesystem.IFileSystem;
 import com.nervousfish.nervousfish.service_locator.IServiceLocator;
